refactor: move LlamaMarketTemplate one level up
