+++ apps/main/src/llamalend/features/borrow/components/BorrowTabContents.tsx
@@ -1,7 +1,7 @@
 import { useCallback, useEffect } from 'react'
 import type { ReactNode } from 'react'
 import { FormProvider } from 'react-hook-form'
-import type { NetworkDict } from '@/llamalend/llamalend.types'
+import type { LlamaMarketTemplate, NetworkDict } from '@/llamalend/llamalend.types'
 import type { IChainId } from '@curvefi/llamalend-api/lib/interfaces'
 import Box from '@mui/material/Box'
 import Button from '@mui/material/Button'
@@ -10,7 +10,7 @@ import Stack from '@mui/material/Stack'
 import { useBorrowPreset } from '@ui-kit/hooks/useLocalStorage'
 import { t } from '@ui-kit/lib/i18n'
 import { SizesAndSpaces } from '@ui-kit/themes/design/1_sizes_spaces'
-import { BorrowPreset, type LlamaMarketTemplate, type OnBorrowFormUpdate } from '../borrow.types'
+import { BorrowPreset, type OnBorrowFormUpdate } from '../borrow.types'
 import { hasLeverage } from '../llama.util'
 import { setValueOptions } from '../llama.util'
 import { useBorrowForm } from '../useBorrowForm'
