+++ apps/main/src/llamalend/features/borrow/llama.util.ts
@@ -1,12 +1,12 @@
 import type { SetValueConfig } from 'react-hook-form'
 import type { Address } from 'viem'
 import { zeroAddress } from 'viem'
+import type { LlamaMarketTemplate } from '@/llamalend/llamalend.types'
 import type { INetworkName } from '@curvefi/llamalend-api/lib/interfaces'
 import { LendMarketTemplate } from '@curvefi/llamalend-api/lib/lendMarkets'
 import { MintMarketTemplate } from '@curvefi/llamalend-api/lib/mintMarkets'
 import { requireLib } from '@ui-kit/features/connect-wallet'
 import { assert, CRVUSD_ADDRESS } from '@ui-kit/utils'
-import { LlamaMarketTemplate } from './borrow.types'
 import { BorrowPreset } from './borrow.types'
 
 /**
