+++ apps/main/src/llamalend/features/borrow/useBorrowForm.tsx
@@ -2,6 +2,7 @@ import { useEffect, useMemo } from 'react'
 import type { UseFormReturn } from 'react-hook-form'
 import { useForm } from 'react-hook-form'
 import { useAccount } from 'wagmi'
+import type { LlamaMarketTemplate } from '@/llamalend/llamalend.types'
 import type { INetworkName } from '@curvefi/llamalend-api/lib/interfaces'
 import type { IChainId } from '@curvefi/llamalend-api/lib/interfaces'
 import { notFalsy, recordEntries } from '@curvefi/prices-api/objects.util'
@@ -10,7 +11,7 @@ import type { BaseConfig } from '@ui/utils'
 import { SLIPPAGE_PRESETS } from '@ui-kit/features/slippage-settings/ui/slippage.utils'
 import { useDebouncedValue } from '@ui-kit/hooks/useDebounce'
 import { formDefaultOptions } from '@ui-kit/lib/model'
-import { type BorrowForm, BorrowPreset, type LlamaMarketTemplate } from './borrow.types'
+import { type BorrowForm, BorrowPreset } from './borrow.types'
 import { useMaxTokenValues } from './hooks/useMaxTokenValues'
 import { getTokens } from './llama.util'
 import { BORROW_PRESET_RANGES } from './llama.util'
